requirements.txt
coverage==4.2
dj-database-url==0.4.1
Django==1.10.1
django-configurations==2.0
django-localflavor==1.3
django-nose==1.4.4
gunicorn==19.6.0
nose==1.3.7
psycopg2==2.6.2
requests==2.11.1
rollbar==0.13.5
six==1.10.0
whitenoise==3.2.1
